fix OpenXmlformats Sheet and Style issues; change HPSF test cases; change all the vs2010 project file to target .net 4.0
git-svn-id: http://npoi.googlecode.com/svn/trunk@460 3d0cb27c-5942-9f1f-b0c8-b51bd5f61ba7

+++ testcases/main/POIFS/NIO/TestDataSource.cs
@@ -89,9 +89,9 @@ namespace TestCases.POIFS.NIO
         public void TestByteArray()
         {
             byte[] data = new byte[256];
-            for (byte i = 0; i < data.Length; i++)
+            for (int i = 0; i < data.Length; i++)
             {
-                data[i] = i;
+                data[i] = (byte)i;
             }
 
             ByteArrayBackedDataSource ds = new ByteArrayBackedDataSource(data);
